Следующий коммит будет нормальный, честно.
Нормальное работает поиск пути с зумом и движением камеры. Также путь ищется с учетом занятых блоков.

+++ src/main/kotlin/world/objects/mobs/Dog.kt
@@ -20,6 +20,7 @@ import ImageHelper.Companion.dog_walk_right_bottom
 import ImageHelper.Companion.dog_walk_right_top
 import ImageHelper.Companion.dog_walk_top
 import objects.Activity
+import world.ChunkAndPoint
 import world.objects.Direction
 import world.objects.mobs.activityBehaviors.ActivityBehavior
 import world.objects.mobs.activityBehaviors.WalkBehavior
@@ -27,7 +28,7 @@ import java.awt.Image
 import java.awt.event.ActionListener
 import javax.swing.Timer
 
-class Dog(chunkAndPoint: ChunkAndPoint, width: Int = 2, height: Int = 4, override var direction: Direction = Direction.BOTTOM,
+class Dog(chunkAndPoint: ChunkAndPoint, width: Int = 2, height: Double = 4.0, override var direction: Direction = Direction.BOTTOM,
           override var activity: Activity = Activity.STAND,
 //             override val stepsForBlock: Int = 8,
           override var step: Int = 0
@@ -49,6 +50,7 @@ class Dog(chunkAndPoint: ChunkAndPoint, width: Int = 2, height: Int = 4, overrid
     }
 
     override var behavior: ActivityBehavior? = ActivityBehaviorHelper.dogBehavior()
+
     override fun getImage(): Image {
         when (activity){
             Activity.STAND ->
