+++ src/main/kotlin/world/objects/mobs/Knight.kt
@@ -19,11 +19,12 @@ import ImageHelper.Companion.knight_standing_right_top
 import ImageHelper.Companion.knight_standing_top
 import Point
 import objects.Activity
+import world.ChunkAndPoint
 import world.objects.Direction
 import world.objects.mobs.activityBehaviors.ActivityBehavior
 import java.awt.Image
 
-class Knight(chunkAndPoint: ChunkAndPoint, width: Int = 2, height: Int = 4, override var direction: Direction = Direction.BOTTOM,
+class Knight(chunkAndPoint: ChunkAndPoint, width: Int = 2, height: Double = 4.0, override var direction: Direction = Direction.BOTTOM,
              override var activity: Activity = Activity.STAND,
 //             override val stepsForBlock: Int = 4,
              override var step: Int = 0
