+++ src/main/kotlin/world/objects/mobs/Peasant.kt
@@ -26,11 +26,12 @@ import ImageHelper.Companion.peasant_walk_right_bottom
 import ImageHelper.Companion.peasant_walk_right_top
 import ImageHelper.Companion.peasant_walk_top
 import objects.Activity
+import world.ChunkAndPoint
 import world.objects.Direction
 import world.objects.mobs.activityBehaviors.ActivityBehavior
 import java.awt.Image
 
-class Peasant(chunkAndPoint: ChunkAndPoint, width: Int = 2, height: Int = 4, override var direction: Direction = Direction.BOTTOM,
+class Peasant(chunkAndPoint: ChunkAndPoint, width: Int = 2, height: Double = 4.0, override var direction: Direction = Direction.BOTTOM,
               override var activity: Activity = Activity.STAND,
 //             override val stepsForBlock: Int = 8,
               override var step: Int = 0
